improve by introducing directory finder and organizing files
