+++ Kununu/ArchitectureTest/Configuration/Selector/InterfaceClassSelector.php
@@ -1,16 +1,16 @@
 <?php
 declare(strict_types=1);
 
-namespace Kununu\ArchitectureTest\Configuration;
+namespace Kununu\ArchitectureTest\Configuration\Selector;
 
 use PHPat\Selector\Selector;
 use PHPat\Selector\SelectorInterface;
 
-final readonly class NamespaceSelector implements Selectable
+final readonly class InterfaceClassSelector implements Selectable
 {
     use RegexTrait;
 
-    public const string KEY = 'namespace';
+    public const string KEY = 'interface';
 
     public function __construct(
         public string $name,
@@ -22,7 +22,10 @@ final readonly class NamespaceSelector implements Selectable
     {
         $namespace = $this->makeRegex($this->namespace);
 
-        return Selector::inNamespace($namespace, $namespace !== $this->namespace);
+        return Selector::AllOf(
+            Selector::classname($namespace, $namespace !== $this->namespace),
+            Selector::isInterface(),
+        );
     }
 
     public function getName(): string
